+++ storybook/src/components/document-card/document-card.content.ts
@@ -1,4 +1,4 @@
-import { InfoButton } from "dso-toolkit";
+import { InfoButton, Label } from "dso-toolkit";
 import { TemplateResult, html } from "lit-html";
 
 import { Templates } from "../../templates";
@@ -6,6 +6,19 @@ import { Templates } from "../../templates";
 export const typeItems = [
   html`<span class="dso-document-card-type-item">Omgevingsplan</span>`,
   html`<span class="dso-document-card-type-item">Gemeente Amsterdam</span>`,
+];
+
+export const labels: Label[] = [
+  {
+    label: "Bouwwerken",
+    compact: true,
+    status: "bright",
+  },
+  {
+    label: "Kleine windturbine",
+    compact: true,
+    status: "bright",
+  },
 ];
 
 export function infoButton({ richContentTemplate }: Templates): InfoButton<TemplateResult> {